Remove the edgesByPortId constructor param from GraphAttribute
